flask_blog/requirements.txt
# basic flask
flask==3.0.0
MarkupSafe==2.1.3
Werkzeug==3.0.1
blinker==1.7.0
click==8.1.7
Jinja2==3.1.2
itsdangerous==2.1.2

# forms
flask-wtf==1.2.1
wtforms==3.1.1
dnspython==2.4.2
email_validator==2.1.0post1
idna==3.4

# database
flask_sqlalchemy==3.1.1
greenlet==3.0.1
sqlalchemy==2.0.23
typing_extensions==4.8.0

# hash password
bcrypt==4.0.1
flask_bcrypt==1.0.1

# login
flask_login==0.6.3

# file resizer
pillow==10.1.0

# authorization (jwt)
PyJWT==2.80

# mail
Flask-Mail==0.9.1
